+++ src/main/java/roomescape/config/CheckAdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.ui;
+package roomescape.config;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -8,21 +8,16 @@ import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.infrastructure.JwtTokenProvider;
 
 @Component
-public class CheckLoginInterceptor implements HandlerInterceptor {
+public class CheckAdminInterceptor implements HandlerInterceptor {
     private final JwtTokenProvider jwtTokenProvider;
 
-    public CheckLoginInterceptor(JwtTokenProvider jwtTokenProvider) {
+    public CheckAdminInterceptor(JwtTokenProvider jwtTokenProvider) {
         this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         Cookie[] cookies = request.getCookies();
-        if (cookies == null) {
-            response.setStatus(401);
-            return false;
-        }
-        jwtTokenProvider.validateToken(cookies);
-        return true;
+        return jwtTokenProvider.checkRoleByCookie(cookies, response);
     }
 }
