+++ src/main/java/roomescape/config/MemberIdConverter.java
@@ -1,4 +1,4 @@
-package roomescape.ui;
+package roomescape.config;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
