+++ src/main/java/roomescape/config/MemberIdConverterArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.ui;
+package roomescape.config;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
