+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -5,9 +5,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.ui.CheckLoginInterceptor;
-import roomescape.ui.CheckAdminInterceptor;
-import roomescape.ui.MemberIdConverterArgumentResolver;
 
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
