+++ src/main/java/roomescape/domain/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
