+++ src/main/java/roomescape/repository/H2MemberRepository.java
@@ -12,6 +12,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Member;
+import roomescape.domain.MemberRepository;
 import roomescape.domain.Role;
 
 @Repository
