+++ src/main/java/roomescape/service/ThemeService.java
@@ -5,8 +5,8 @@ import roomescape.dto.request.ThemeRequest;
 import roomescape.dto.response.ThemeResponse;
 import roomescape.domain.Reservation;
 import roomescape.domain.Theme;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ThemeRepository;
+import roomescape.domain.ReservationRepository;
+import roomescape.domain.ThemeRepository;
 
 import java.time.LocalDate;
 import java.util.Comparator;
